refactor: group components by section
