+++ components/about/mission.tsx
