+++ components/events/event-list-view.tsx
@@ -1,11 +1,11 @@
 import Link from "next/link";
 import { BiTimeFive } from "react-icons/bi";
 import { GrLocation } from "react-icons/gr";
-import { eventColors } from "../lib/eventColors";
-import sameDay from "../lib/sameDay";
-import { EventInCMS } from "../types/event-in-cms";
-import FormattedDate from "./formatted-date";
-import FormattedTime from "./formatted-time";
+import { eventColors } from "../../lib/eventColors";
+import sameDay from "../../lib/sameDay";
+import { EventInCMS } from "../../types/event-in-cms";
+import FormattedDate from "../formatted-date";
+import FormattedTime from "../formatted-time";
 
 interface Props {
   eventsArray: EventInCMS[];
