+++ pages/about.tsx
@@ -1,7 +1,7 @@
 import Head from "next/head";
-import Description from "../components/description";
-import Mission from "../components/mission";
-import Vision from "../components/vision";
+import Description from "../components/about/description";
+import Mission from "../components/about/mission";
+import Vision from "../components/about/vision";
 
 export default function About() {
   return (
