+++ pages/events.tsx
@@ -2,9 +2,9 @@ import { useState } from "react";
 import Modal from "react-modal";
 import { getEvents } from "../lib/posts";
 import markdownToHtml from "../lib/markdownToHtml";
-import EventsGrid from "../components/events-grid";
-import EventFeatured from "../components/event-featured";
-import EventShowcase from "../components/event-showcase";
+import EventsGrid from "../components/events/events-grid";
+import EventFeatured from "../components/events/event-featured";
+import EventShowcase from "../components/events/event-showcase";
 import Head from "next/head";
 import { EventInCMS, isArrayOfEventsInCMS } from "../types/event-in-cms";
 import { IoClose } from "react-icons/io5";
